requirements.txt
appdirs==1.4.3
atari-py==0.0.19
box2d-py==2.3.1
cycler==0.10.0
dask==0.14.1
decorator==4.0.11
doom-py==0.0.15
gym==0.8.1
imageio==2.1.2
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.2.0
jupyter-client==5.0.0
jupyter-core==4.3.0
Keras==2.0.2
matplotlib==2.0.0
mujoco-py==0.5.7
networkx==1.11
numpy==1.12.1
olefile==0.44
pachi-py==0.0.21
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
ppaquette-gym-doom==0.0.6
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyglet==1.2.4
Pygments==2.2.0
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
requests==2.13.0
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
Theano==0.9.0
toolz==0.8.2
torch==0.1.10.post2
torchvision==0.1.7
tornado==4.4.2
traitlets==4.3.2
vizdoom==1.1.1
wcwidth==0.1.7
